% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lncRNAanalysis.R
\name{Coexpression}
\alias{Coexpression}
\title{Analysis of lncRNA co expression with mRNA}
\usage{
Coexpression(lncname, pcname, normData)
}
\arguments{
\item{lncname}{a vector of Ensembl lncRNA id}

\item{pcname}{a vector of Ensembl protein coding gene id}

\item{normData}{normlized Expression matrix after log Transformation,by \code{\link{DEAnalysis}}}
}
\value{
a datafram of correlation analysis
}
\description{
compute correlation between lncRNA with mRNA
}
