% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools.R
\name{CombinePlot}
\alias{CombinePlot}
\title{Target prediction}
\usage{
CombinePlot(plots, ncol = NULL, legend = NULL, ...)
}
\arguments{
\item{plots}{A list of gg objects}

\item{ncol}{Number of columns}

\item{legend}{Combine legends into a single legend choose
from 'right' or 'bottom'; pass 'none' to remove 
legends, or NULL to leave legends as they are}

\item{...}{Extra parameters passed to plot_grid}
}
\description{
Combine ggplot2-based plots into a single plot,copy from Seurat
}
\author{
zexl
}
