% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseqAnalysis.R
\name{DEAnalysis}
\alias{DEAnalysis}
\title{Routine analysis of transcriptome sequencing}
\usage{
DEAnalysis(countDataRaw, Group, control, filter = TRUE, lognormdata = TRUE)
}
\arguments{
\item{countDataRaw}{Expression matrix of transcriptome sequencing}

\item{Group}{a vector of sample group}

\item{control}{Set up control group}

\item{filter}{a logical ,filter gene use cpm}

\item{lognormdata}{a logical,rlogTransformation for normdata}
}
\value{
a list,data frame of Difference analysis,normlized Expression matrix ,
 and normlized Expression matrix after log Transformation
}
\description{
Difference analysis of transcriptome sequencing
}
\references{
Robinson MD, McCarthy DJ, Smyth GK. edgeR: a Bioconductor package
    for differential expression analysis of digital gene expression data.
    Bioinformatics. 2010 Jan 1;26(1):139-40. \cr
    Ritchie ME, Phipson B, Wu D, Hu Y, Law CW, Shi W, Smyth GK.
    limma powers differential expression analyses for RNA-sequencing and
    microarray studies. Nucleic acids research. 2015 Jan 20;
    43(7):e47-e47. \cr
    Love MI, Huber W, Anders S. Moderated estimation of fold change and
    dispersion for RNA-seq data with DESeq2. Genome biology. 2014 Dec 5;
    15(12):550.
    @author zexl
    @examples
    DeResult<-DEAnalysis(Counts,
            group,control='Z')
    DEGAll<-DeResult[[1]]
    normData<-DeResult[[2]]
    normData_log <- assay(
       rlogTransformation(DeResult[[3]])
       )
}
\author{
zexl
}
