% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceRNAAnalysis.R
\name{ExportNetwork}
\alias{ExportNetwork}
\title{Export network for Cytoscape}
\usage{
ExportNetwork(ceNetwork, net)
}
\arguments{
\item{ceNetwork}{a dataframe generated from \code{\link{ceRNAAnslysis}}}

\item{net}{one of \code{'nodes'} and \code{'edges'}}
}
\value{
A dataframe of nodes or edges
}
\description{
Export nodes and edges of ce network for 
    \strong{Cytoscape} visualization
}
\author{
zexl
}
