% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseqAnalysis.R
\name{GOEnrich}
\alias{GOEnrich}
\title{GO enrichment analysis}
\usage{
GOEnrich(gene, simplify = TRUE, level = 0, data.type = "")
}
\arguments{
\item{gene}{a vector of Ensembl gene id}

\item{simplify}{logical, specifying whether to remove redundant GO terms.
Default simplify=TRUE}

\item{level}{a numeric value, restrict the GO enrichment result at a
specific GO level. Default is \code{0}, which means all terms
should be returned}

\item{data.type}{prefix for Save file names}
}
\value{
return a data frame of GO enrichment analysis result, and write it to file
}
\description{
Performs Gene Ontology (GO) enrichment
    analyses by clusterProfiler packages
}
\examples{
    Upgene<-deall[which(deall$logFC>=1),]
    UpgeneGoOut<-GOEnrich(
            gene = rownames(Upgene),
            data.type = 'RNA_Up'
            )
}
\references{
Yu G, Wang LG, Han Y, He QY. clusterProfiler: an R package for
    comparing biological themes among gene clusters.
    Omics: a journal of integrative biology. 2012 May 1;16(5):284-7.

Yu G, Wang LG, Yan GR, He QY. DOSE: an R/Bioconductor package for
    disease ontology semantic and enrichment analysis. Bioinformatics.
    2014 Oct 17;31(4):608-9.
}
\author{
zexl
}
