% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enrichmentplot.R
\name{GOplots}
\alias{GOplots}
\title{Plots for enrichment analysis}
\usage{
GOplots(
  enrichment,
  type = "bar",
  num.terms = 10,
  axis = "all",
  data.type = "",
  w = 0,
  h = 0,
  save = ""
)
}
\arguments{
\item{enrichment}{a dataframe generated from
\code{\link{GOEnrich}}}

\item{type}{type of the plot, should be one of \code{'bar'}
and \code{'bubble'}}

\item{num.terms}{number of terms to be plotted. Default is \code{10}}

\item{axis}{y axis,GO number \code{term},GO Description \code{description},
or GO number~Description \code{all},Default is \code{all}}

\item{data.type}{prefix for Save file names}

\item{w}{image weight}

\item{h}{image height}

\item{save}{Output file type,default is not save}
}
\description{
Bar plot and bubble plot for GO enrichment analysis
}
