% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseqPlot.R
\name{HeatmapPlot}
\alias{HeatmapPlot}
\title{heatmap}
\usage{
HeatmapPlot(exprdata, Group = group, datatype = "", save = "")
}
\arguments{
\item{exprdata}{a dataframe for plot heatmap}

\item{Group}{sample group}

\item{datatype}{prefix for Save file names}

\item{save}{Output file type,default is not save}
}
\description{
A heatmap showing unsupervised hierarchical clustering of
    DE genes/miRNAs
}
\author{
zexl
}
