% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseqAnalysis.R
\name{KEGGEnrich}
\alias{KEGGEnrich}
\title{KEGG Enrichment}
\usage{
KEGGEnrich(gene, data.type = "")
}
\arguments{
\item{gene}{a vector of Ensembl gene id}

\item{data.type}{prefix for Save file names}
}
\value{
return a data frame of KEGG enrichment analysis result, and write it to file
}
\description{
Performs GeneKyoto Encyclopedia of Genes and Genomes KEGG
    pathway enrichment analyses by clusterProfiler packages
}
\examples{
    Upgene<-deall[which(deall$logFC>=1),]
    UpgeneKEGGOut<-KEGGEnrich(
           gene = rownames(Upgene),
           data.type = 'RNA_Up'
           )
}
\references{
Yu G, Wang LG, Han Y, He QY. clusterProfiler: an R package for
    comparing biological themes among gene clusters.
    Omics: a journal of integrative biology. 2012 May 1;16(5):284-7. \cr
    Yu G, Wang LG, Yan GR, He QY. DOSE: an R/Bioconductor package for
    disease ontology semantic and enrichment analysis. Bioinformatics.
    2014 Oct 17;31(4):608-9.
}
\author{
zexl
}
