% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enrichmentplot.R
\name{KEGGplots}
\alias{KEGGplots}
\title{Plots for enrichment analysis}
\usage{
KEGGplots(
  kegg,
  type = "bar",
  num.terms = 10,
  w = 0,
  h = 0,
  data.type = "",
  save = ""
)
}
\arguments{
\item{kegg}{a dataframe generated from
\code{\link{KEGGEnrich}} or \code{\link{DOEnrich}}}

\item{type}{type of the plot, should be one of \code{'bar'}
and \code{'bubble'}}

\item{num.terms}{number of terms to be plotted. Default is \code{10}}

\item{w}{image weight}

\item{h}{image height}

\item{data.type}{prefix for Save file names}

\item{save}{Output file type,default is not save}
}
\description{
Bar plot and bubble plot for KEGG/DO enrichment analysis
}
