% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lncRNAanalysis.R
\name{mergeCorForNeargene}
\alias{mergeCorForNeargene}
\title{lncRNA cis}
\usage{
mergeCorForNeargene(correlationresult)
}
\arguments{
\item{correlationresult}{a datafram of correlation analysis}
}
\value{
a datafram of lncRNA cis analysis results
}
\description{
Calculate the distance between lncrna and its adjacent coding genes,
 and combine the correlation results
}
