% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseqPlot.R
\name{normPlot}
\alias{normPlot}
\title{Combination of box line diagram and density diagram}
\usage{
normPlot(countDataRaw, exprSet, datatype)
}
\arguments{
\item{countDataRaw}{Expression data}

\item{exprSet}{data after normlizing}

\item{datatype}{prefix for Save file names}
}
\description{
Box line diagram and density diagram before and after standardization
}
\author{
zexl
}
