% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools.R
\name{outline}
\alias{outline}
\title{Remove outliers}
\usage{
outline(datavector, qnum = 0.75)
}
\arguments{
\item{datavector}{target vector}

\item{qnum}{Specify quantile for separation outliers,default is 0.75}
}
\value{
Vectors with outliers process
}
\description{
separation vector outliers replace outliers
}
\author{
zexl
}
