% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseqPlot.R
\name{PCA_GGplot}
\alias{PCA_GGplot}
\title{PCA plot}
\usage{
PCA_GGplot(rnaExpr = rnaExpr, group, datatype = "", save = "")
}
\arguments{
\item{rnaExpr}{a matrix ,Expression data after standardization}

\item{group}{Sample grouping}

\item{datatype}{prefix for Save file names}

\item{save}{Output file type,default is not save}
}
\description{
Principal component analysis and visualization
}
\author{
zexl

zexl
}
