% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools.R
\name{plot_3dPCA}
\alias{plot_3dPCA}
\title{3D PCA}
\usage{
plot_3dPCA(exprdata, Group, title = "3dPCA")
}
\arguments{
\item{exprdata}{matrix for PCA}

\item{Group}{samples group}

\item{title}{Title of PCA map}
}
\description{
PCA analysis and plot 3D PCA
}
\author{
zexl
}
