% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lncRNAplot.R
\name{plotcis}
\alias{plotcis}
\title{Plots for lncRNA cis analysis}
\usage{
plotcis(NearPlotData, data.dir = ".", save = "")
}
\arguments{
\item{NearPlotData}{a data frame of lncRNA cis analysis}

\item{data.dir}{Specify save path,default is Current working path}

\item{Output}{file type,default is not save}
}
\description{
Plots for lncRNA cis analysis,lnc-pc cor pvalue distance data visualization
}
\author{
zexl
}
