% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools.R
\name{rocplot}
\alias{rocplot}
\title{ROC curve}
\usage{
rocplot(rdata, data.type = "", save = "")
}
\arguments{
\item{rdata}{pROC results}

\item{data.type}{prefix for Save file names}

\item{save}{Output file type,default is not save}
}
\description{
ROC curve drawing
}
\author{
zexl
}
