% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lncRNAanalysis.R
\name{TFanalysis.Fisher}
\alias{TFanalysis.Fisher}
\title{Correlation analysis of transcription factors}
\usage{
TFanalysis.Fisher(test.frame, prex)
}
\arguments{
\item{test.frame}{Correlation analysis results of designated lncrna}

\item{prex}{prefix for Save file names}
}
\description{
Based on the results of co expression, Fisher's exact
 test was used to analyze the transcription factor association of single lncrna
}
\author{
zexl
}
