% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseqPlot.R
\name{VolcanoPlot}
\alias{VolcanoPlot}
\title{volcano plot}
\usage{
VolcanoPlot(
  deg.all,
  fc = 2,
  pval = 0.05,
  data.type = "mRNA",
  nlab = 0,
  save = ""
)
}
\arguments{
\item{deg.all}{a dataframe generated from  \code{\link{DEAnalysis}}
containing all genes of analysis no matter they are differentially
expressed or not}

\item{fc}{a numeric value specifying the threshold of fold change}

\item{pval}{a nuemric value specifying the threshold of p value
default is 0.05}

\item{data.type}{prefix for Save file names}

\item{nlab}{a nuemric value,Specify the number of genes to display the gene name}

\item{save}{Output file type,default is not save}
}
\description{
A volcano plot showing differentially expressed genes/miRNAs
}
\author{
zexl
}
