% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools.R
\name{writeDEdata}
\alias{writeDEdata}
\title{write difference analysis results to csv}
\usage{
writeDEdata(wfile, c1name = "", outfileprex = "", Counts = NULL)
}
\arguments{
\item{wfile}{a datafram need to write}

\item{c1name}{The row name is the gene name, the row name is write to 
the first column, and the column name is specified}

\item{outfileprex}{prefix for Save file names}

\item{Counts}{counts matrix,normdata result of DEAnalysis function, 
or raw count matrix}
}
\description{
Write the result of differential analysis to 
    CSV to avoid excel turning on gene name into month,
     and combine the result of differential analysis with counts
}
\author{
zexl
}
