% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools.R
\name{WriteTargets}
\alias{WriteTargets}
\title{Target prediction}
\usage{
WriteTargets(miRNA, database = "starBase", datatype = "Gene")
}
\arguments{
\item{miRNA}{miRNAs}

\item{database}{miRNA target genes prediction data base,
default is 'starBase',or 'mirTarBase' 'miRcode'}

\item{datatype}{prefix for Save file names}
}
\description{
predict miRNA target genes
}
\author{
zexl
}
